requirements.txt
pywhatkit>=5.4
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
schedule>=1.2.0
numpy>=1.24.0
